Add all initial file for project
