+++ lib/modules/welcome_screen.dart
@@ -1,0 +1,22 @@
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+class WelcomeScreen extends StatelessWidget {
+  const WelcomeScreen({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: const Text('Quiz App'),
+        centerTitle: true,
+        backgroundColor: Colors.lightBlue,
+      ),
+      body: Center(
+        child: Column(
+          children: [Text('Testing')],
+        ),
+      ),
+    );
+  }
+}
